chore(tests): run all cloud tests in parallel.
