+++ cloud/testserver/users.go
@@ -1,0 +1,22 @@
+// Copyright 2023 Terramate GmbH
+// SPDX-License-Identifier: MPL-2.0
+
+package testserver
+
+import "net/http"
+
+type userHandler struct{}
+
+func newUserEndpoint() *userHandler {
+	return &userHandler{}
+}
+
+func (userHandler *userHandler) ServeHTTP(w http.ResponseWriter, _ *http.Request) {
+	w.Header().Add("Content-Type", "application/json")
+	writeString(w, `{
+			    "email": "batman@example.com",
+			    "display_name": "batman",
+				"job_title": "entrepreneur"
+			}`,
+	)
+}
