+++ cmd/terramate/e2etests/logging_test.go
@@ -10,6 +10,7 @@ import (
 )
 
 func TestLoggingChangeChannel(t *testing.T) {
+	t.Parallel()
 	s := sandbox.New(t)
 	cli := newCLI(t, s.RootDir())
 	cli.loglevel = "trace"
