+++ test/cloud/client.go
@@ -13,18 +13,16 @@ import (
 	"github.com/terramate-io/terramate/cloud"
 )
 
-// TestEndpoint is the localhost endpoint used by tests.
-const TestEndpoint = "http://localhost:3001"
 const defaultTestTimeout = 1 * time.Second
 
 // PutStack sets a new stack in the /v1/stacks/<org>/<stack id>.
 // Note: this is not a real endpoint.
-func PutStack(t *testing.T, orgUUID string, st cloud.StackResponse) {
+func PutStack(t *testing.T, addr string, orgUUID string, st cloud.StackResponse) {
 	ctx, cancel := context.WithTimeout(context.Background(), defaultTestTimeout)
 	defer cancel()
 
 	client := &cloud.Client{
-		BaseURL:    TestEndpoint,
+		BaseURL:    "http://" + addr,
 		Credential: &credential{},
 	}
 	_, err := cloud.Put[cloud.EmptyResponse](ctx, client, st, cloud.StacksPath, orgUUID, strconv.Itoa(st.ID))
